refactor: 데이터 수집 모듈 분리 #19
